requirements.txt
streamlit
streamlit-extras
huggingface_hub
transformers
torch
faiss-cpu
sentence-transformers
google-generativeai
Pillow
psutil
rapidfuzz

